move function to `snap-utils` and add JSDocs
