+++ packages/snaps-utils/src/index.ts
@@ -1,4 +1,5 @@
 export type * from './account';
+export * from './address';
 export * from './array';
 export * from './auxiliary-files';
 export * from './base64';
